feat(common): 페이지 템플릿을 common 으로 이동, ErrorPage 추가 외
- env 타입명에 대한 오타 수정
